+++ envisim_utils/tests/searcher.rs
@@ -1,6 +1,7 @@
 use envisim_test_utils::*;
 use envisim_utils::kd_tree::*;
 use envisim_utils::{Matrix, Probabilities};
+use std::num::NonZeroUsize;
 
 fn matrix_new<'a>() -> Matrix<'a> {
     Matrix::new(
@@ -19,12 +20,12 @@ fn searcher() -> Result<(), NodeError> {
         .try_bucket_size(2)?
         .build(&mut [0, 1, 2, 3])?;
 
-    let mut s = Searcher::new(&t, 1).unwrap();
+    let mut s = Searcher::new_1(&t);
     s.find_neighbours(&t, &vec![5.0, 5.0]).unwrap();
     assert_eq!(s.neighbours(), vec![1]);
     assert_delta!(s.distance_k(0), 41.0);
 
-    let mut s = Searcher::new(&t, 2).unwrap();
+    let mut s = Searcher::new(&t, NonZeroUsize::new(2).unwrap());
     s.find_neighbours_of_id(&t, 3).unwrap();
     assert_eq!(s.neighbours(), vec![2, 1]);
     assert_delta!(s.distance_k(0), 221.0);
@@ -41,7 +42,7 @@ fn searcher_weighted() -> Result<(), NodeError> {
         .build(&mut [0, 1, 2, 3, 4])?;
     let p = Probabilities::new(5, 0.25).unwrap();
 
-    let mut s = SearcherWeighted::new(&t).unwrap();
+    let mut s = SearcherWeighted::new(&t);
     s.find_neighbours(&t, &p, &vec![5.0, 5.0], 0.5).unwrap();
     assert_eq!(s.neighbours(), vec![1, 0]);
     assert_delta!(s.weight_k(0), 0.5);