+++ src/utils.rs
@@ -10,8 +10,8 @@
 // You should have received a copy of the GNU Affero General Public License along with this
 // program. If not, see <https://www.gnu.org/licenses/>.
 
-use crate::SampleOptions;
-use envisim_utils::{Indices, Probabilities, SamplingError};
+use crate::{SampleOptions, SamplingError};
+use envisim_utils::{Indices, Probabilities};
 use rand::Rng;
 
 pub struct Sample(Vec<usize>);