+++ envisim_estimate/CHANGELOG.md
@@ -7,5 +7,16 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.2.0] - 2024-09-24
+### Added
+- added dependency `envisim_samplr`.
+
+### Changed
+- `n_neighbours` parameter of `local_mean_variance` changed type from `usize` to `NonZeroUsize`.
+
+### Removed
+- removed dependency `thiserror`.
+- removed re-exports from `spatial_balance` module.
+
 ## [0.1.0] - 2024-09-19
 Initial release.